Fixed a crash, increased slide audio pitch
Game would crash when moving a piece off the right side of the board.

Changed audio frequency to be more similar to that of a real chess piece on a real chess board. Still not accurate, of course. The sound is just white noise and all I can change is the frequency.